added blog controller

+++ Program.cs
@@ -5,7 +5,7 @@ var builder = WebApplication.CreateBuilder(args);
 
 builder.WebHost.ConfigureKestrel(option =>
 {
-    option.Limits.MaxRequestBodySize = int.MaxValue; // Setting MaxRequestBodySize to a large value
+    option.Limits.MaxRequestBodySize = long.MaxValue; // Setting MaxRequestBodySize to a large value
 });
 
 // Add services to the container.
@@ -14,7 +14,7 @@ builder.Services.AddControllersWithViews();
 // Configure FormOptions to handle large file uploads
 builder.Services.Configure<FormOptions>(options =>
 {
-    options.MultipartBodyLengthLimit = int.MaxValue; // Setting MultipartBodyLengthLimit to a large value
+    options.MultipartBodyLengthLimit = long.MaxValue; // Setting MultipartBodyLengthLimit to a large value
 });
 
 var config = builder.Configuration;
